Changed genes related to water pathing
Partially because of my own ineptitude and partially because of my own preference, I removed most movespeed offsets from genes and plugged in code used in 1.6 update Webbed Phalanges in order to make pathing work properly. The end speeds seems to be similar to one that was in previous updates, though I haven't played extensively back then so I cannot be completely certain. Full list of changes:
- Ear fins and pelvic fins are now purely cosmetic genes. I removed water speed offset and complexity costs.
- Webbed hands are now a purely negative trait with no water speed offset.
- Aquatic is now a purely negative trait, whose only purpose is to cause Hydration need and be a prerequisite gene for the tail.
- Tail now has the "waterCellCost 1" tag, exactly as 1.6upd Webbed phalanges, which eliminates all pathing costs from moving over water. I tried to see if I can change the value to make them move even quicker in water than on land, but that didn't seem to work.
Swift swim is left unchanged for now. It works alright with merfolks, but I imagine that if it's injected separately from the tail the pathfinding problem will return.